Add code for releases 2.2, 2.3, 2.4, 2.5

+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apiindicator/model/EquipmentImpactBO.java
@@ -4,7 +4,7 @@
  *
  * This product includes software developed by
  * French Ecological Ministery (https://gitlab-forge.din.developpement-durable.gouv.fr/pub/numeco/m4g/numecoeval)
- */ 
+ */
 package com.soprasteria.g4it.backend.apiindicator.model;
 
 import lombok.AllArgsConstructor;
@@ -28,8 +28,8 @@ public class EquipmentImpactBO {
 
     private String status;
 
-    private Float impact;
+    private Double impact;
 
-    private Float sip;
+    private Double sip;
 
 }
